requirements.txt
google-generativeai==0.8.3
requests==2.32.3
streamlit==1.41.1
beautifulsoup4==4.12.3
trafilatura==1.6.1
python-dotenv==1.0.0
tiktoken==0.5.1
readability-lxml==0.8.1